requirements.txt
attrs==17.4.0
pluggy==0.6.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
py==1.5.2
pytest==3.4.0
python-dateutil==2.6.1
sh==1.12.14
six==1.11.0
z3c.schema==1.1.0
zope.event==4.3.0
zope.i18nmessageid==4.1.0
zope.interface==4.4.3
zope.schema==4.5.0
